chore: moved dashboard specific components to dashboard folder under components for consistency
