+++ commitment/imports/ui/components/dashboard/BookmarkFilter.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Filter, { FilterOption, FiltersState, FilterValue } from '../../ui/filter'
+import Filter, { FilterOption, FiltersState, FilterValue } from '../ui/filter'
 
 
 interface FilterProps {
